Quan fe

+++ BE_DuAn/src/main/java/com/poly/be_duan/service/DesignService.java
@@ -10,9 +10,9 @@ public interface DesignService {
 
     public Designs create(Designs designs);
 
-    public Optional<Designs> getDesignByID(Long id);
+    public Optional<Designs> getDesignByID(Integer id);
 
     public Designs update(Designs designs);
 
-    public void delete(Long id);
+    public void delete(Integer id);
 }
